refactor: complete App.tsx break up
