+++ packages/app/src/components/Nav/NavLogo.tsx
@@ -4,7 +4,7 @@ function NavLogo(): JSX.Element {
     return (
         <div className="flex flex-col items-end">
             <h3 className="text-3xl font-bold uppercase hover:text-white">Orion</h3>
-            <div className="-mt-1 -mr-2 px-2 text-lg text-white font-semibold bg-purple-400 rounded-full">
+            <div className="-mr-2 px-2 text-lg text-white font-semibold bg-purple-400 rounded-full">
                 labs
             </div>
         </div>
